backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langfuse>=2.0.0
python-dotenv